PyDocPlugin:
Added `sys.path` configuration setting for `[pydoc]`.

The Python documentation will only be displayed for builtin modules and for
the modules listed there (use the `os.pathsep` character in order to separate
the paths).

If the configuration key is not set, the `sys.path` variable value will be 
used (as it is now).

git-svn-id: http://trac-hacks.org/svn@309 7322e99d-02ea-0310-aa39-e9a107903beb