TimingAndEstimationPlugin:
trying to get postgres to work during trac upgrades.  It was giving errors about aborted transactions.  In playing in IPython, I was able to figure out that it had somethign to do with needing to commit after every use of the connection.  Normally I don't commit after selects.  This has changed. 
see ticket:957

git-svn-id: http://trac-hacks.org/svn@1626 7322e99d-02ea-0310-aa39-e9a107903beb