AddCommentMacro: Warning + `req.redirect()` issues fixed.
The macro used the first version warning style (`req.warning()`), but that was just a left-over from before moving to redirect on successful submit anyway - it would not have displayed to the user no matter what. The line is just removed. Fixes #2368.

The whole `req.redirect()` idea needed to be reworked as macros cannot really do redirects. A redirect raises `RequestDone`, and like all other exceptions from macros it gets swallowed by the Formatter - causing an `AssertionError` to appear in the logs when the request tries to write a second status/header/response output later.

The exception did not display to the user, but was annoying... Now the redirect sets a marker on the `req` object, and by implementing `IRequestFilter` it now re-raises the `RequestDone` there.


git-svn-id: http://trac-hacks.org/svn@3027 7322e99d-02ea-0310-aa39-e9a107903beb