closes #2761
Sorry for the continuing errors.  Please upgrade again (from svn up is fine, if you still have the source control copy available).

There was recently a large refactoring on this very young plugin, and while I thought it was pretty fully working, apparently I missed a few code paths related to the upgrade process.

Hopefully this will be the last of your troubles

git-svn-id: http://trac-hacks.org/svn@3404 7322e99d-02ea-0310-aa39-e9a107903beb