re #3209
Based on what you linked to, this seems like a problem using the str() function when I should be using unicode()

The place that I found where it was doing this was in building up some json output.  I changed this to unicode, which will hopefully mean that this error is solved.


HTH,

Russ

git-svn-id: http://trac-hacks.org/svn@4340 7322e99d-02ea-0310-aa39-e9a107903beb