Alter the way to list up system pages (#3832).
Now we know system page names by listing files in 'wiki-default'
directory. This is a solution of for ticket #3832.

Note that the pages provided by plugins (they are not placed in
wiki-default directory) are not categorized as system page any more.

There is no way to force them categorizing as system page yet,
but it may be implemented in near future.


git-svn-id: http://trac-hacks.org/svn@4367 7322e99d-02ea-0310-aa39-e9a107903beb