WikiTicketCalendarMacro: Add a third set of backward/forward navigation links, some code cleanup.
This is the last backwards-kompatible change introduced with introduced by SVN r7893 before.
It fills the gap between jumping just one month and one year quite nicely.
I'm still not satisfied with the optic, but this could be easily fixed, 
if someone comes up with a visually more attractiv and - most important - 
more intuitive looking choice of charakters for all link labels.
The experiment with temporarily resetting locale settings, 
that was left from tests for the unicode fix in previous version, is removed.


git-svn-id: http://trac-hacks.org/svn@8054 7322e99d-02ea-0310-aa39-e9a107903beb