WikiTicketCalendarMacro: Finally update trunk to latest development version.
This is the 1st version compatible with Trac 0.12, fixing issues with 
the change in time stamp format to POSIX microseconds with Trac 0.12dev-r9210.
Code is essentially the same as committed in SVN r7893 before, but we 
have a history and three branches now.
To show this remarkable step I declare the trunk to be version 1.0.0.
But that's just another start. Let's go on from here.


git-svn-id: http://trac-hacks.org/svn@8056 7322e99d-02ea-0310-aa39-e9a107903beb