new functionality: add the parameter {{{showaggregatedstate}}} to the renderer {{{tableticketperuserday}}}
while setting {{{showaggregatedstate}}} to {{{True}}}, {{{T}}}, {{{Y}}}, or {{{yes}}} the following functionality is initiated: all table cells containing a ticket are jugded by a red, blue, or green bar. This indicates whether the tickets are untouched, in work, or finished, i.e., the states of all tickets terminated at the considered date are aggregated.

 * precondition: the considered tickets are terminated before the current date
 * note: currently the states of the following workflows are considered (http://trac.edgewall.org/wiki/WorkFlow/Examples):
   * original Trac 0.10 workflow
   * enterprise workflow



git-svn-id: http://trac-hacks.org/svn@9880 7322e99d-02ea-0310-aa39-e9a107903beb