TracFormsPlugin: Fix errors related to special resources and encoding.
I've seen form contexts like '/newticket', that will not validate to a
known Trac realm, and throw an Error
 ''locale variable 'subcontext' referenced before assignment''
instead because of misplacement of corresponding variable declaration.
Good news is, new db table definition do prevent recording of form data in
such strange (pseudo-)resources effectively in the future, but surely
the code might profit from a more verbose error message, when trying so.

The strftime format suddenly happend to be unicode on a production system,
so I fixed this as well. Apart from this the code looks quite usable so far
and arbitrary upgrades should be safe to follow. I'll provide all follow-ups
from this point of development onwards.

git-svn-id: http://trac-hacks.org/svn@10005 7322e99d-02ea-0310-aa39-e9a107903beb