TracFormsPlugin: Make stepwise history reset method available, refs #3445 and #4759.
I've made a decision to not support an single form field reset. So with this
changeset the !TracForm reset support is nearly final as far as I consider
it to be necessary, and reasonably usable according to my tests.
But a form reset is a final decision, there is even no confirmation dialog,
so `FORM_RESET` permission should be granted with care to keep this actions
administrative.
Again: Think __before__ pressing the almighty button(s) - you've been warned.

git-svn-id: http://trac-hacks.org/svn@10127 7322e99d-02ea-0310-aa39-e9a107903beb