WikiCalendarMacro: Merge in WikiCalendarMacro code for a combined release, refs #7564.
This is by no means an atomic changeset but a huge, crazy code clash.
Both macros had rather different approaches regarding class definitions for
CSS styles, so I had to partially change core logic to allow the creation of
some `WikiCalendar` specific HTML tags. While the CSS changes are not final,
the first goal of a regression-free product is reached, or at least close
enough for the visual presentation of both macros.

Short summary of changes:
 * include full embedded macro documentation for `WikiCalendar`
 * use 3-section-navigation from `WikiTicketCalendar` with both calendars now
 * combine CSS style definitions
  * remove dedicated subfolder 'css' from htdocs - too much for just on file
  * move CSS namespace `wikiTicketCalendar` to `wikitcalendar`
  * reuse/merge identical/similar definitions, unify CSS class names (started)
 * some lessons learned from `WikiCalendar` styling
  * move formatting from HTML to CSS (started, day numbers and subpages)
  * use multiple classes per HTML tag (started with day numbers)
 * fix an overwrite of marco name in milestone code

git-svn-id: http://trac-hacks.org/svn@10203 7322e99d-02ea-0310-aa39-e9a107903beb