Revert "Make default due date handling consistent.  Refs #9443."
This reverts commit a16d6abc70b1ab724f5a0236625440ce00a4602d.

This introduces a bug.  To say nothing of the fact that it doesn't seem
to address the problem.  While the placement of the milestone diamond is
wrong, when I view page source the time of the milestone start and
finish is the end of the day so the problem is in the drawing.

git-svn-id: http://trac-hacks.org/svn@10875 7322e99d-02ea-0310-aa39-e9a107903beb