AnnouncerPlugin: Revert [9114] for breaking Trac 0.11 and 0.12 compatibility, refs #9522.
Actually `ChoiceOption` appeared even after Trac 0.12 release, so this would
belong into a dedicated `0.13` branch, that we may have in the future.

I plan to rollback and convert any 0.12 stuff to regain backwards-compatibility
with Trac down to 0.11, and move development to 0.13 instead.  This involves
mostly i18n stuff (could be done in a more compatible fashion too) and
especially the db connection handling, that has seen may changes.
It seems to have settled only lately in Trac 0.13dev .

git-svn-id: http://trac-hacks.org/svn@10919 7322e99d-02ea-0310-aa39-e9a107903beb