Allow configuration of normal work week. Refs #9735.
The original plugin assumed if there was no data in the database, that
no one was working.

With this change, that is still the default, but you can configure
work_days in the team-calendar section of trac.ini to a list of week
days that people usually work and that will be used if no data is in the
database.  0 is Monday so 0,1,2,3,4 is a "normal" Monday-Friday week
with Saturday and Sunday off.

git-svn-id: http://trac-hacks.org/svn@11207 7322e99d-02ea-0310-aa39-e9a107903beb