TracAnnouncer: Extend changelog with plugin history for better overview, refs #3578.
This is rather complicated due to branching without version syncronisation.

I draw the conclusion, that moving to an unambiguous version number requires
overpowering `1.0-a1`, so I'm pushing up to the lowest satisfying number now.
`announcer-1.0` will support any Trac release down to and including 0.11, if
feasible at all, so the match with Trac-1.0 is rather incidential this time.

git-svn-id: http://trac-hacks.org/svn@12224 7322e99d-02ea-0310-aa39-e9a107903beb