Refactor query.  Refs #9648.
Instead of having PM-aware applications (like the Gantt) do

 * Use TracPM to preprocess options
 * Use Trac's TicketQuery()
 * Use TracPM to postprocess results

We augment TracPM to provide a `query()` function which encapsulates
those steps.  The return value is now also pruned so there are no
"dangling" references to external tickets.  If A is a parent of B is
a parent of C and you query `root=B`, the list is cleaned up so asking
for the parent of B says there isn't one.

As a result, we can clean up some Gantt options and code

 * root= and goal= aren't really Gantt options, they are query options supported by TracPM.
 * omitMilestones isn't really a TracPM query option
 * The Gantt doesn't have to jump through hoops to tell if there is a parent ticket

Also, this allows us to reuse, rather than copy, the query logic within
TracPM.

(Fix a typo in pmapi.py while I'm there.)

The now-private preQuery() returns a set, not a list, now.

git-svn-id: http://trac-hacks.org/svn@12329 7322e99d-02ea-0310-aa39-e9a107903beb