Assign WBS in correct order.  Refs #10868, #9648.
The order of WBS assignments was broken by r12314.

`TracPM.roots()` returns an unordered list of ticket IDs so traversing
that list meant we were no longer computing the WBS in the correct
order.  Now we use that list to determine what is a root ticket but
traverse the ticket list in the scheduled (sorted) order.

Thanks to falkb for his help and patience while working on this.

git-svn-id: http://trac-hacks.org/svn@12426 7322e99d-02ea-0310-aa39-e9a107903beb