Fixed issues with [13312].
Rename of a variable in [13312] (`dbver` -> `db_version`) caused a conflict with another variable of the same name in the scope, and would have caused future database upgrades to fail.

In [13312] the plugin name was changed from `ExtendedVersionTracPlugin` to `ExtendedVersionPlugin`. The references to the plugin name in the logger have been updated now as well.

git-svn-id: http://trac-hacks.org/svn@13317 7322e99d-02ea-0310-aa39-e9a107903beb