Complete rework of plugin, using ideas from #11210 in the patch by chris-at. Fixes #7554, Refs #8821, #11210.
Note that only one instance of the report in `to_recipients` is removed. So, for instance, if the reporter is also the owner, and `always_notify_owner` is `True`, then the reporter will still receive a notification.

This has been lightly tested, so please report back if you find any issues.

git-svn-id: http://trac-hacks.org/svn@13701 7322e99d-02ea-0310-aa39-e9a107903beb