Don't consider group start, finish when resource leveling.  Refs #9648.
Consider one group and separate task.  All tasks are one unit of work.

 * Group A is owned by Mickey
 * Tasks B, C, and D are children of A. All are owned by Minnie
 * Task E is owned by Mickey and has no dependencies.

An ASAP scheduler would try to schedule, which requires scheduling B,
then C, then D.  When resource leveling, they will happen one after the
other.  The way the existing algorithm bubbles up finish dates, Group A
gets a finish of 3 and that is assigned to the group owner, Mickey.

Now the scheduler tries to schedule E.  Checking the resource limits, it
thinks Mickey isn't available until 3 and scheduled E from 3 to 4,
leaving Mickey idle for three units at the start of the schedule.

Since there is no scheduled work in a group, the owner's availability is
not affected by the work done in the children.  This change skips
consideration and update resource limits if a task has children.

git-svn-id: http://trac-hacks.org/svn@13812 7322e99d-02ea-0310-aa39-e9a107903beb