DiscussionPlugin: Clean-up for environment setup participant, refs #6783.
Most notably passing a db cursor outside of the scope of the constructing
method or function shall be avoided, because it causes rather nasty,
hard-to-debug issues. Rephrasing 'db' to 'schema' for clarity.

We do not handle exceptions on db upgrade but let Trac deal with it instead,
enabling it to better care itself for side-effects on other setup participants.
Introducing some INFO logging gives a bit more insight behind the scene of
running upgrade actions.

git-svn-id: http://trac-hacks.org/svn@13887 7322e99d-02ea-0310-aa39-e9a107903beb