3.0.0dev: Log warning only if component exists
The purpose of the warning message is to help locate components that correspond to non-existent projects. If the component and wiki page have been removed, then there is no need to log a warning. The latter can occur when visiting a ticket that has the component field set to a project that has been removed.

Refs #12252.


git-svn-id: https://trac-hacks.org/svn@14965 7322e99d-02ea-0310-aa39-e9a107903beb