11.0.0.6: Skip permission check on resource containing the `TOC`
Permission must have already been granted on this resource
if the macro is being executed. An effect of this change is that
`TOC` can be used in resources other than wiki pages, such as
blog posts, as long as the `TOC` is only created for the post
containing the `TOC` macro.

Fixes #12914.


git-svn-id: https://trac-hacks.org/svn@15937 7322e99d-02ea-0310-aa39-e9a107903beb