Allow escaping characters using backslashes
A backslash can be inserted in order to escape a character in the macro
parameters and give it its literal meaning. Typically this would be a comma (to
include the comma in the substitution rather than using it to separate multiple
parameters). Backslashes themselves can be escaped with more backslashes.


git-svn-id: https://trac-hacks.org/svn@17349 7322e99d-02ea-0310-aa39-e9a107903beb